+++ CHANGELOG.md
@@ -580,3 +580,8 @@ Breaking Changes:
 * Roles are replaced by "Presets" and "Permission" keys. Read more [here](https://coralogix.com/docs/api-keys/).
 
 Various documentation upgrades
+
+## Release 1.16.1
+New Features:
+#### resource/coralogix_alert
+* adding `more_than_or_equal_usual` and `less_than_or_equal_usual` conditions to `metric.promql` alert.